refactor: extract sampling functions to _sampling.py
